separate PostHogLogLevel and PostHogLogSeverity
